confirmar exclusão

+++ src/app/pages/categories/category-list/cotegory-list.component.ts
@@ -22,10 +22,13 @@ export class CategoryListComponent implements OnInit {
     );
   }
   deleteCategory(category: CategoryModel) {
-    this.categoryService.deleteCategory(category.id).subscribe(() => {
-     this.categories =  this.categories.filter(element => element != category);
-      alert("Deletado com sucesso!")
-    }, () => alert("Erro ao tentar excluir."));
-
+    const mustDelete = confirm("Deseja realmente excluir este item?");
+    if(mustDelete){
+      this.categoryService.deleteCategory(category.id).subscribe(() => {
+       this.categories =  this.categories.filter(element => element != category);
+        alert("Deletado com sucesso!")
+      }, () => alert("Erro ao tentar excluir."));
+    }
   }
+  
 }
